feat: modernization epic — tooling gaps + structural debt (prexu-nlqf) (#14)
Closes epic prexu-nlqf (9/9 tasks).

- ci: oxlint + clippy `-D warnings` gates, lint scripts, .editorconfig;
fixed 20 pre-existing clippy findings
- api: Zod 4 validation at Plex service boundary (safeParse + graceful
fallback), removed 26 unsafe casts
- build(ts): tsconfig ES2022 + noUncheckedIndexedAccess +
verbatimModuleSyntax (148 fallout fixes)
- build(vite): enable React Compiler (babel-plugin-react-compiler 1.0.0)
- data: adopt TanStack Query; hooks replace hand-rolled
fetch/cancel/chunk loops
- refactor(pages): DetailPageShell collapses Collection/Playlist
duplication (1826→1604 LOC)
- refactor(player): extract player/lifecycle.rs; perf(relay): broadcast
Arc<String>
- logging: bare console.* → logger service
- build(rust): pin toolchain 1.94.1 + .nvmrc

Gates: tsc 0, oxlint 0, vitest 1718 pass, clippy clean both crates, prod
build ok.

Manual smoke needed (behavioral): React Compiler, Collection/Playlist
detail pages (live data fetching, play/edit/delete/reorder/group,
empty+error states).

+++ src-tauri/src/player/taskbar_preview.rs
@@ -16,6 +16,7 @@
 //! only while mpv is alive), and answer DWM's pull requests:
 //!   - `WM_DWMSENDICONICTHUMBNAIL`   → one scaled frame (alt-tab tile).
 //!   - `WM_DWMSENDICONICLIVEPREVIEW` → a frame for the taskbar/Aero-Peek hover.
+//!
 //! A ~10 Hz `DwmInvalidateIconicBitmaps` heartbeat (only while playing) makes
 //! the live preview update so the user can watch by hovering, like VLC/Plex.
 //! Invalidation is pull-based: DWM only re-requests a bitmap while one is
@@ -278,7 +279,7 @@ unsafe fn make_dib(w: i32, h: i32) -> Option<(HBITMAP, *mut u8)> {
             biHeight: -h, // negative => top-down, matching screenshot-raw rows
             biPlanes: 1,
             biBitCount: 32,
-            biCompression: BI_RGB.0 as u32,
+            biCompression: BI_RGB.0,
             ..Default::default()
         },
         ..Default::default()
